+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Druid/GroveShaper.cs
@@ -3,10 +3,12 @@ using HearthDb.Enums;
 
 namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem.Cards.Druid;
 
-public class GroveShaper : ICardWithHighlight
+public class GroveShaper : ICardWithHighlight, ISpellSchoolTutor
 {
 	public string GetCardId() => HearthDb.CardIds.Collectible.Druid.GroveShaper;
 
 	public HighlightColor ShouldHighlight(Card card, IEnumerable<Card> deck) =>
-		HighlightColorHelper.GetHighlightColor(card.GetTag(GameTag.SPELL_SCHOOL)  == (int)SpellSchool.NATURE);
+		HighlightColorHelper.GetHighlightColor(card.GetTag(GameTag.SPELL_SCHOOL) == (int)SpellSchool.NATURE);
+
+	public int[] TutoredSpellSchools { get; } = { (int)SpellSchool.NATURE };
 }